+++ k8s1.15.0/node_install_k8s.sh
@@ -1,18 +1,18 @@
 #!/bin/bash
 #b8_yang@163.com
-source ./base.config
 bash_path=$(cd "$(dirname "$0")";pwd)
+source $bash_path/base.config
 
 if [[ "$(whoami)" != "root" ]]; then
 	echo "please run this script as root ." >&2
 	exit 1
 fi
 
-log="./setup.log"  #操作日志存放路径 
-fsize=2000000         
+log="./setup.log"  #操作日志存放路径
+fsize=2000000
 exec 2>>$log  #如果执行过程中有错误信息均输出到日志文件中
 
-echo -e "\033[31m 这个是centos7系统初始化脚本，请慎重运行！Please continue to enter or ctrl+C to cancel \033[0m"
+echo -e "\033[31m 这个是k8s一键部署脚本，node节点正在运行脚本中,请不要刷新或断开连接，结束会有相关提示！Please continue to enter or ctrl+C to cancel \033[0m"
 #sleep 5
 #yum update
 yum_update(){
@@ -21,27 +21,58 @@ yum_update(){
 #configure yum source
 yum_config(){
   yum install wget epel-release -y
-  cd /etc/yum.repos.d/ && mkdir bak && mv -f *.repo bak/
-  wget -O /etc/yum.repos.d/CentOS-Base.repo http://mirrors.aliyun.com/repo/Centos-7.repo
-  wget -O /etc/yum.repos.d/epel.repo http://mirrors.aliyun.com/repo/epel-7.repo
-  yum clean all && yum makecache
-  yum -y install iotop iftop yum-utils net-tools git lrzsz expect gcc gcc-c++ make cmake libxml2-devel openssl-devel curl curl-devel unzip sudo ntp libaio-devel wget vim ncurses-devel autoconf automake zlib-devel  python-devel bash-completion
-#  ntpdate 0.asia.pool.ntp.org
+  if [[ $aliyun == "1" ]];then
+  test -d /etc/yum.repos.d/bak/ || yum install wget epel-release -y && cd /etc/yum.repos.d/ && mkdir bak && mv -f *.repo bak/ && wget -O /etc/yum.repos.d/CentOS-Base.repo http://mirrors.aliyun.com/repo/Centos-7.repo && wget -O /etc/yum.repos.d/epel.repo http://mirrors.aliyun.com/repo/epel-7.repo && yum clean all && yum makecache
+  fi
 }
+
+yum_init(){
+num=0
+while true ; do
+let num+=1
+yum -y install iotop iftop yum-utils net-tools git lrzsz expect gcc gcc-c++ make cmake libxml2-devel openssl-devel curl curl-devel unzip sudo ntp libaio-devel wget vim ncurses-devel autoconf automake zlib-devel  python-devel bash-completion
+if [[ $? -eq 0 ]] ; then
+echo "初始化安装环境配置完成！！！"
+break;
+else
+if [[ num -gt 3 ]];then
+echo "你登录 "$masterip" 瞅瞅咋回事？一直无法yum包"
+break
+fi
+echo "FK!~没成功？哥再来一次！！"
+fi
+done
+}
+
 #firewalld
 iptables_config(){
+if [[ `ps -ef | grep firewalld |wc -l` -gt 1 ]];then
   systemctl stop firewalld.service
   systemctl disable firewalld.service
-  iptables -P FORWARD ACCEPT
+  echo "防火墙我关了奥！！！"
+fi
+#  iptables -P FORWARD ACCEPT
 }
+
 #system config
 system_config(){
+grep "SELINUX=disabled" /etc/selinux/config
+if [[ $? -eq 0 ]];then
+  echo "SELINUX 已经禁用！！"
+else
   sed -i "s/SELINUX=enforcing/SELINUX=disabled/g" /etc/selinux/config
   setenforce 0
-  timedatectl set-local-rtc 1 && timedatectl set-timezone Asia/Shanghai
-#  yum -y install chrony && systemctl start chronyd.service && systemctl enable chronyd.service
+  echo "SELINUX 已经禁用！！"
+fi
+
 }
+
+
 ulimit_config(){
+grep 'ulimit' /etc/rc.local
+if [[ $? -eq 0 ]];then
+echo "内核参数调整完毕！！！"
+else
   echo "ulimit -SHn 102400" >> /etc/rc.local
   cat >> /etc/security/limits.conf << EOF
   *           soft   nofile       102400
@@ -51,17 +82,24 @@ ulimit_config(){
   *           soft  memlock      unlimited 
   *           hard  memlock      unlimited
 EOF
-
+  cat >> /etc/sysctl.conf << EOF
+    kernel.pid_max=4194303
+EOF
+sysctl -p
+echo "内核参数调整完毕！！！"
+fi
 }
 
 ssh_config(){
-
-if [`grep 'UserKnownHostsFile' /etc/ssh/ssh_config`];then
-echo "pass"
+grep 'UserKnownHostsFile' /etc/ssh/ssh_config
+if [[ $? -eq 0 ]];then
+echo "ssh参数配置完毕！！！"
 else
 sed -i "2i StrictHostKeyChecking no\nUserKnownHostsFile /dev/null" /etc/ssh/ssh_config
+echo "ssh参数配置完毕！！！"
 fi
 }
+
 
 #set sysctl
 sysctl_config(){
@@ -71,7 +109,6 @@ sysctl_config(){
   net.bridge.bridge-nf-call-ip6tables = 1
   net.ipv4.ip_nonlocal_bind = 1
   net.ipv4.ip_forward = 1
-  vm.swappiness=0
 EOF
   /sbin/sysctl -p
   echo "sysctl set OK!!"
@@ -79,14 +116,18 @@ EOF
 
 #swapoff
 swapoff(){
+grep 'vm.swappiness=0' /etc/sysctl.conf
+if [[ $? -eq 0 ]];then
+echo "临时命名空间删除！！！"
+else
   /sbin/swapoff -a
   sed -i '/ swap / s/^\(.*\)$/#\1/g' /etc/fstab
   echo "vm.swappiness=0" >> /etc/sysctl.conf
   /sbin/sysctl -p
+fi
 }
 
 get_localip(){
-ipaddr='172.0.0.1'
 ipaddr=$(ip addr | awk '/^[0-9]+: / {}; /inet.*global/ {print gensub(/(.*)\/(.*)/, "\\1", "g", $2)}' | grep $ip_segment)
 echo "$ipaddr"
 }
@@ -101,18 +142,17 @@ setupkernel(){
 
 #ssh trust
 rootssh_trust(){
-rm -rf ~/.ssh
 cd $bash_path
 for host in ${hostip[@]}
 do
 if [[ `get_localip` != $host ]];then
-#ls /root/.ssh
+
 if [[ ! -f /root/.ssh/id_rsa.pub ]];then
 expect ssh_trust_init.exp $root_passwd $host
 else
 expect ssh_trust_add.exp $root_passwd $host
 fi
-echo "remote machine root user succeed!!!!!!!!!!!!!!!! "
+echo "服务器互信完成！！！ "
 fi
 done
 }
@@ -124,13 +164,45 @@ echo $hash_value
 
 #install docker
 install_docker() {
+test -d /etc/docker
+if [[ $? -eq 0 ]];then
+echo "docker已经安装完毕!!!"
+else
+mkdir -p /etc/docker
 yum-config-manager --add-repo  https://mirrors.aliyun.com/docker-ce/linux/centos/docker-ce.repo
 yum install -y --setopt=obsoletes=0 docker-ce-18.09.4-3.el7
-systemctl start docker
+tee /etc/docker/daemon.json <<-'EOF'
+{
+  "registry-mirrors": ["https://gpkhi0nk.mirror.aliyuncs.com"]
+}
+EOF
+systemctl daemon-reload
 systemctl enable docker
+systemctl restart docker
+echo "docker已经安装完毕!!!"
+fi
 }
 
+
+# config docker
+config_docker(){
+grep "tcp://0.0.0.0:2375" /usr/lib/systemd/system/docker.service
+if [[ $? -eq 0 ]];then
+echo "docker API接口已经配置完毕"
+else
+sed -i "/^ExecStart/cExecStart=\/usr\/bin\/dockerd -H tcp:\/\/0\.0\.0\.0:2375 -H unix:\/\/\/var\/run\/docker.sock" /usr/lib/systemd/system/docker.service
+systemctl daemon-reload
+systemctl restart docker.service
+echo "docker API接口已经配置完毕"
+fi
+}
+
+
 set_repo(){
+test -f /etc/yum.repos.d/kubernetes.repo
+if [[ $? -eq 0 ]];then
+echo "kubelet kubectl kubeadm安装完毕！！!"
+else
 cat <<EOF > /etc/yum.repos.d/kubernetes.repo
 [kubernetes]
 name=Kubernetes
@@ -145,6 +217,7 @@ EOF
 	systemctl daemon-reload
 	systemctl enable kubelet
 	systemctl start kubelet
+fi
 }
 
 install_masterk8s(){
@@ -164,13 +237,13 @@ install_masterk8s(){
 }
 
 install_flannel(){
-	cd $bash_path
-	wget https://raw.githubusercontent.com/coreos/flannel/bc79dd1505b0c8681ece4de4c0d86c5cd2643275/Documentation/kube-flannel.yml
+    cd $bash_path
+    test -f kube-flannel.yml || wget https://raw.githubusercontent.com/coreos/flannel/bc79dd1505b0c8681ece4de4c0d86c5cd2643275/Documentation/kube-flannel.yml
 	kubectl apply -f kube-flannel.yml
+    echo "flannel 网络配置完毕"
 }
 
 join_cluster(){
-
  kubeadm join --token $tocken --discovery-token-ca-cert-hash sha256:$sha_value $masterip:6443
 }
 
@@ -178,23 +251,22 @@ join_cluster(){
 
 main(){
  #yum_update
- #setupkernel
  yum_config
+ yum_init
  ssh_config
  iptables_config
  system_config
- #ulimit_config
  sysctl_config
-# change_hosts
+
  swapoff
  install_docker
+ #config_docker
  set_repo
- 
  rootssh_trust
- 
  install_masterk8s
  install_flannel
  
  join_cluster
+ echo `hostname`" 已经成功加入k8s集群"
 }
 main > ./setup.log 2>&1
